+++ test/site.sh
@@ -28,3 +28,8 @@ IPv6=true
 # This check is optional because some installs, such as vagrant make a non-systemd/direct
 # start
 NGINXCHECK=true
+
+# Lenght of TSL RSA key
+# Problem with small key such as 1024 not allowed in centos8 for example (why is this)
+# Problem with long keys are they take time to generate, eg on ARM
+CERTKEYLEN=2048